Skill_Matrix_AI/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.3.2
chromadb==0.4.22
pypdf==3.17.4
langchain==0.1.12
torch==2.2.1
numpy==1.26.4
langchain-google-genai==0.0.5
langchain-community==0.0.29
python-multipart==0.0.9
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0
sentence-transformers==2.2.2
transformers==4.30.2
huggingface-hub==0.16.4
pathlib==1.0.1
PyPDF2==3.0.1
